+++ packages/studiocms/src/cli/cmds/add/askToContinue.ts
@@ -1,0 +1,14 @@
+import * as p from '@clack/prompts';
+
+export async function askToContinue(): Promise<boolean> {
+	const response = await p.confirm({
+		message: 'Continue?',
+		initialValue: true,
+	});
+
+	if (p.isCancel(response)) {
+		return false;
+	}
+
+	return response;
+}
